DockerPrueba/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6  # <-- ¡Añade esta línea!
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2